requirements.txt
# Core
torch>=2.2.0
transformers>=4.44.0
tokenizers>=0.19.0
accelerate>=0.33.0
safetensors>=0.4.0

# Data & ML
scikit-learn>=1.5.0
numpy>=1.26.0
pandas>=2.2.0

# Sentence embeddings for persona classifier
sentence-transformers>=3.0.0

# Evaluation (Ollama via REST API)
requests>=2.31.0

# Experiment tracking
wandb>=0.17.0

# Config
pyyaml>=6.0

# Progress & logging
tqdm>=4.66.0

# Visualization (for analysis phase)
matplotlib>=3.9.0
seaborn>=0.13.0
